feat(button): fix sonar issues

+++ game/src/exitlayer.cpp
@@ -56,7 +56,7 @@ bool ExitLayer::onUpdate(uint32_t elapsedTime) {
 
     std::string_view message = "Exit the Game?";
     uint32_t length = font->getLength(message, 48);
-    font->render(message, { (width - length) / 2.F, height / 2.F - 128.F }, 48,
+    font->render(message, { (width - static_cast<float>(length)) / 2.F, height / 2.F - 128.F }, 48,
                  { 1.F, 1.F, 1.F });
 
     confirmButton->onUpdate(elapsedTime);
